backend/requirements.txt
asgiref==3.8.1
cffi==2.0.0
chardet==5.2.0
cryptography==46.0.2
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
sqlparse==0.5.3
tzdata==2025.2
Werkzeug==3.1.3
